// src/modules/webhooks/enums/webhook-category.enum.ts
export enum WebhookCategory {
  CardTransaction = 'card_transaction',
  CardAuthTransaction = 'card_auth_transaction',
  CardFeePatch = 'card_fee_patch',
  Card3ds = 'card_3ds',
  CardHolder = 'card_holder',
  PhysicalCard = 'physical_card',
  Work = 'work',
  WalletTransaction = 'wallet_transaction',
}